fix: Add more tests

+++ storage3/_async/resumable.py
@@ -5,7 +5,7 @@ from datetime import datetime
 from ..types import FileInfo, UploadMetadata
 from ..utils import AsyncClient, FileStore, StorageException
 
-__all__ = ("AsyncResumableUpload",)
+__all__ = ["AsyncResumableUpload"]
 
 
 class AsyncResumableUpload:
@@ -15,11 +15,9 @@ class AsyncResumableUpload:
         self.expiration_time_format = "%a, %d %b %Y %X %Z"
         self._filestore = FileStore()
 
-    def _is_valid_arg(self, target):
-        return isinstance(target, str) and len(target.strip()) > 0
-
     def _encode(self, metadata: UploadMetadata) -> str:
         """Generate base64 encoding for Upload-Metadata header
+
         Parameters
         ----------
         metadata
@@ -30,30 +28,26 @@ class AsyncResumableUpload:
         ]
         return ",".join(res)
 
-    def file_exists(self, filename) -> bool:
-        """Verify if the file exists in the storage
-        Parameters
-        ----------
-        filename
-            This could be the local filename or objectname in the storage
-        """
-        return filename in self._filestore.storage
-
     def get_link(self, objectname) -> str:
         """Get the link associated with objectname in the bucket
+
         Parameters
         ----------
         objectname
             This could be the local filename or objectname in the storage
         """
-        if not self.file_exists(objectname):
-            raise StorageException(f"There is no entry for {objectname} in FileStore")
+        if not self.is_valid_arg(objectname):
+            raise StorageException("bucketname cannot be empty")
         return self._filestore.get_link(objectname)
+
+    def is_valid_arg(self, target):
+        return target is not None and isinstance(target, str) and len(target.strip()) != 0
 
     async def create_unique_link(
         self, bucketname=None, objectname=None, filename=None
     ) -> None:
         """Create unique link according to bucketname and objectname
+
         Parameters
         ----------
         bucketname
@@ -63,17 +57,19 @@ class AsyncResumableUpload:
         filename
             Local file
         """
-        if not self._is_valid_arg(bucketname):
-            raise StorageException("Bucketname cannot be empty")
+        if not self.is_valid_arg(bucketname):
+            raise StorageException("bucketname cannot be empty")
 
-        if not (self._is_valid_arg(objectname) or self._is_valid_arg(filename)):
+        if not (self.is_valid_arg(objectname) or self.is_valid_arg(filename)):
             raise StorageException("Must specify objectname or filename")
 
         file = filename if filename else objectname
-        if not self._is_valid_arg(file):
+
+        if not self.is_valid_arg(file):
             raise StorageException("Must specify objectname or filename")
 
         upload_mode = None
+
         info = FileInfo(
             name=file, link="", length="", headers={"Tus-Resumable": "1.0.0"}
         )
@@ -113,6 +109,7 @@ class AsyncResumableUpload:
 
     async def resumable_offset(self, link, headers) -> str:
         """Get the current offset to be used
+
         Parameters
         ----------
         link
@@ -120,7 +117,15 @@ class AsyncResumableUpload:
         headers
             Metadata headers sent to the server
         """
+
+        if not self._filestore.link_exists(link):
+            raise StorageException(f"There's no a reference to that link: {link}")
+
         response = await self._client.head(link, headers=headers)
+
+        if "upload-offset" not in response.headers:
+            raise StorageException("Error while fetching the next offset.")
+
         return response.headers["upload-offset"]
 
     async def terminate(self, file: str) -> None:
@@ -131,6 +136,9 @@ class AsyncResumableUpload:
         file
             file name used to get its metadata info
         """
+        if not self.is_valid_arg(file):
+            raise StorageException("file argument cannot be empty")
+
         info = self._filestore.get_file_info(file)
         response = await self._client.delete(info["link"], headers=info["headers"])
 
@@ -143,6 +151,7 @@ class AsyncResumableUpload:
         self, filename, upload_defer=False, link=None, objectname=None, mb_size=1
     ) -> None:
         """Send file's content in chunks to the target url
+
         Parameters
         ----------
         filename
@@ -157,21 +166,21 @@ class AsyncResumableUpload:
             Amount of megabytes to be sent in each iteration
         """
         if upload_defer:
-            if not (self._is_valid_arg(link) and self._is_valid_arg(objectname)):
+            if not (self.is_valid_arg(link) and self.is_valid_arg(objectname)):
                 raise StorageException(
                     "Upload-Defer mode requires a link and objectname"
                 )
 
-        if not self._is_valid_arg(filename):
+        if not self.is_valid_arg(filename):
             raise StorageException("Must specify a filename")
 
         target_file = objectname if upload_defer else filename
-        chunk_size = 1048576 * int(abs(mb_size))  # 1024 * 1024 * mb_size
+        chunk_size = 1048576 * mb_size
         size = None
         self._filestore.update_file_headers(
             target_file, "Content-Type", "application/offset+octet-stream"
         )
-        storage_link = link if upload_defer else self._filestore.get_link(target_file)
+        storage_link = link if upload_defer else self.get_link(target_file)
 
         if upload_defer:
             size = str(os.stat(filename).st_size)
@@ -193,9 +202,11 @@ class AsyncResumableUpload:
 
             chunk = file.read(chunk_size)
             headers = self._filestore.get_file_headers(target_file)
+
             response = await self._client.patch(
                 storage_link, headers=headers, content=chunk
             )
+
             if response.status_code not in {201, 204}:
                 raise StorageException(response.content)
